requirements.txt
openai
requests
python-dotenv
Pillow
vercel-blob 